[modify] アプリがバックグラウンド状態時，強制終了しないようにデータベース生成をFragment側（TitleFragment）で行うよう修正
