+++ app/src/main/java/com/example/TitleFragment.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,9 +13,11 @@ import android.widget.Button;
 
 import com.example.camera.R;
 import com.example.database.DBHelper;
+import com.example.database.Table;
 import com.example.storyboard.NewStoryFragment;
 import com.example.storyboard.StoryBoardsFragment;
 
+import static com.example.database.DBHelper.TAG;
 
 
 public class TitleFragment extends Fragment {
@@ -33,8 +36,7 @@ public class TitleFragment extends Fragment {
     @Override
     public void onCreate(Bundle saveInstanceState){
         super.onCreate(saveInstanceState);
-        Bundle args = getArguments();
-        dbHelper = (DBHelper) args.getSerializable("DBHelper");
+        createDataBase();
         setRetainInstance(true);
     }
 
@@ -117,6 +119,26 @@ public class TitleFragment extends Fragment {
     };
     */
 
+    public void createDataBase(){
+        try {
+            dbHelper = new DBHelper(getActivity().getApplicationContext());
+            if(dbHelper.getNumOfRecord(Table.COMPOSITION) == -1
+                    || dbHelper.getNumOfRecord(Table.COMPOSITION) == 0) {
+                dbHelper.setComposition();
+            }
+        } catch(Exception e) {
+            Log.d(TAG, e.getMessage());
+        }
+    }
 
+    public void deleteDataBase(){
+        try {
+            dbHelper = new DBHelper(getActivity().getApplicationContext());
+            boolean result = dbHelper.isDatabaseDelete(getActivity().getApplicationContext());
+            Log.d(TAG, " delete result : " + result);
+        } catch(Exception e) {
+            Log.d(TAG, e.getMessage());
+        }
+    }
 
 }
